Refactor + Cleaning code
