+++ server/src/linter.ts
@@ -14,10 +14,8 @@ import { reportValidationMessage } from './functionalities/reporter'
 import { updateDocumentSettings } from './settings'
 import { TimeMeasurer } from './time-measurer'
 import {
-  cursorNode,
-  trimIn,
+  cursorNode, isNodeURI, relativeFilePath, trimIn,
 } from './utils/text-documents'
-import { isNodeURI, relativeFilePath } from './utils/vm/wollok'
 
 // ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
 // INTERNAL FUNCTIONS
