+++ server/src/utils/text-documents.ts
@@ -1,7 +1,8 @@
+import fs from 'fs'
+import path from 'path'
 import { Location, Position, Range, TextDocumentIdentifier, TextDocumentPositionParams } from 'vscode-languageserver'
 import { TextDocument } from 'vscode-languageserver-textdocument'
-import { Environment, Node, Package, SourceIndex, SourceMap } from 'wollok-ts'
-import { relativeFilePath, uriFromRelativeFilePath } from './vm/wollok'
+import { Environment, FileContent, Node, Package, SourceIndex, SourceMap } from 'wollok-ts'
 
 // TODO: Refactor
 const include = (node: Node, { position, textDocument: { uri } }: TextDocumentPositionParams) => {
@@ -9,7 +10,7 @@ const include = (node: Node, { position, textDocument: { uri } }: TextDocumentPo
   if (node.kind === 'Package') {
     return uri.includes(node.sourceFileName)
   }
-  if(!node.sourceMap) return false
+  if (!node.sourceMap) return false
 
   const startPosition = toVSCPosition(node.sourceMap.start)
   const endPosition = toVSCPosition(node.sourceMap.end)
@@ -24,16 +25,16 @@ export const between = (pointer: Position, start: Position, end: Position): bool
   const { line: lineStart, character: charStart } = start
   const { line: lineEnd, character: charEnd } = end
 
-  if(lineStart === lineEnd && linePointer === lineStart) {
+  if (lineStart === lineEnd && linePointer === lineStart) {
     return charPointer >= charStart && charPointer <= charEnd
   }
 
   return linePointer > lineStart
     && linePointer < lineEnd
     || (linePointer === lineStart
-    && charPointer >= charStart
-    || linePointer === lineEnd
-    && charPointer <= charEnd)
+      && charPointer >= charStart
+      || linePointer === lineEnd
+      && charPointer <= charEnd)
 }
 
 export const getNodesByPosition = (environment: Environment, textDocumentPosition: TextDocumentPositionParams): Node[] => {
@@ -53,7 +54,7 @@ export const toVSCRange = (sourceMap: SourceMap): Range =>
   Range.create(toVSCPosition(sourceMap.start), toVSCPosition(sourceMap.end))
 
 export const nodeToLocation = (node: Node): Location => {
-  if(!node.sourceMap || !node.sourceFileName){
+  if (!node.sourceMap || !node.sourceFileName) {
     throw new Error('No source map found for node')
   }
 
@@ -84,9 +85,9 @@ export const packageFromURI = (uri: string, environment: Environment): Package |
 
 export const getWollokFileExtension = (uri: string): 'wlk' | 'wpgm' | 'wtest' => {
   const extension = uri.split('.').pop()
-  if(!extension) throw new Error('Could not determine file extension')
+  if (!extension) throw new Error('Could not determine file extension')
 
-  switch(extension) {
+  switch (extension) {
     case 'wlk':
     case 'wpgm':
     case 'wtest':
@@ -104,4 +105,37 @@ export function cursorNode(
     position,
     textDocument,
   }).reverse()[0]
+}
+
+/** URI */
+
+export let WORKSPACE_URI = ''
+
+export const setWorkspaceUri = (uri: string): void => {
+  WORKSPACE_URI = uri
+}
+
+export const relativeFilePath = (absoluteURI: string): string => {
+  return absoluteURI.replaceAll(WORKSPACE_URI + '/', '')
+}
+
+export const uriFromRelativeFilePath = (relativeURI: string): string => {
+  return WORKSPACE_URI + '/' + relativeURI
+}
+
+export const documentToFile = (doc: TextDocument): FileContent => ({
+  name: relativeFilePath(doc.uri),
+  content: doc.getText(),
+})
+
+export const isNodeURI = (node: Node, uri: string): boolean => node.sourceFileName == relativeFilePath(uri)
+
+export const findPackageJSON = (uri: string): string => {
+  let baseUri = uri
+  while (!fs.existsSync(baseUri + path.sep + 'package.json') && baseUri) {
+    const lastIndex = baseUri.lastIndexOf(path.sep)
+    if (!lastIndex) return ''
+    baseUri = baseUri.slice(0, lastIndex)
+  }
+  return baseUri
 }
